Support Google Play Services fused location backend
